+++ src/errors.rs
@@ -136,6 +136,9 @@ pub enum Error {
     #[error("Unexpected file {path:?} in archive directory")]
     UnexpectedFile { path: String },
 
+    #[error("This feature is not implemented")]
+    NotImplemented,
+
     /// Generic IO error.
     #[error(transparent)]
     IOError {
