+++ src/lib.rs
@@ -27,6 +27,7 @@ pub mod entry;
 pub mod errors;
 pub mod excludes;
 mod gc_lock;
+mod hunk_index;
 pub mod index;
 mod io;
 mod jsonio;
@@ -35,7 +36,6 @@ pub mod live_tree;
 mod merge;
 pub mod misc;
 pub mod monitor;
-#[cfg(windows)]
 mod mount;
 pub mod owner;
 pub mod restore;
@@ -71,8 +71,7 @@ pub use crate::kind::Kind;
 pub use crate::live_tree::LiveTree;
 pub use crate::merge::MergeTrees;
 pub use crate::misc::bytes_to_human_mb;
-#[cfg(windows)]
-pub use crate::mount::mount;
+pub use crate::mount::{mount, MountOptions};
 pub use crate::owner::Owner;
 pub use crate::restore::{restore, RestoreOptions};
 pub use crate::show::{show_versions, ShowVersionsOptions};
